Backend/requirements.txt
# ─── Web Framework ─────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ─── Database ──────────────────────────────────────────────
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
# psycopg2-binary — use pre-built wheel for Windows
psycopg2-binary==2.9.9; sys_platform != "win32"
psycopg2==2.9.9; sys_platform == "win32"

# ─── Task Queue ────────────────────────────────────────────
celery[redis]==5.4.0
redis==5.0.4

# ─── Config ────────────────────────────────────────────────
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1

# ─── HTTP Clients ──────────────────────────────────────────
httpx==0.27.0
aiohttp==3.9.5

# ─── GitHub / AI ───────────────────────────────────────────
anthropic==0.28.0
PyGithub==2.3.0

# ─── Web Auditing ──────────────────────────────────────────
playwright==1.44.0

# ─── Static Analysis (Python tools) ───────────────────────
radon==6.0.1
pylint==3.2.2

# ─── Utilities ─────────────────────────────────────────────
gitpython==3.1.43
python-slugify==8.0.4
tenacity==8.3.0
structlog==24.2.0

# ─── Testing ───────────────────────────────────────────────
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
anyio[trio]==4.3.0
